Replaced classes and rename classes
